Fixed char bug

+++ Core/codeGen.fs
@@ -99,7 +99,7 @@ module CodeGen =
             | Asm(_, asm) -> this.EmitAsm context asm
             | Bool(_, true) -> g.Emit(OpCodes.Ldc_I4_1)
             | Bool(_, false) -> g.Emit(OpCodes.Ldc_I4_0)
-            | Char(_, c) -> g.Emit(OpCodes.Ldc_I4, int16 c)
+            | Char(_, c) -> g.Emit(OpCodes.Ldc_I4, int32 c)
             | Float(_, n) -> g.Emit(OpCodes.Ldc_R4, n)
             | Int(_, n) -> g.Emit(OpCodes.Ldc_I4, n)
             | LookupArg(_, _, i) -> g.Emit(OpCodes.Ldarg, i)
